Avanzando

+++ Auto.cs
@@ -8,7 +8,7 @@ namespace Ordinario_RACA
 {
     internal class Auto : Vehiculo
     {
-        public Auto(int iD, string marca, string modelo, int año, string color,int precio, string estado)
+        public Auto(int iD, string marca, string modelo, int año, string color,double precio, string estado)
         :base (iD, marca, modelo, año, color)
         {
             Precio = precio;
@@ -17,7 +17,7 @@ namespace Ordinario_RACA
 
         
 
-        public int Precio {  get; set; }
+        public double Precio {  get; set; }
         public string Estado { get; set; }
     }
 }
